// packages/frontend/src/services/redux/types.ts
/* eslint-disable no-unused-vars */
export enum ActionType {
  LOGIN_POST = "LOGIN_POST",
  LOGIN_POST_ERROR = "LOGIN_POST_ERROR",
  REGISTER_POST = "REGISTER_POST",
  REGISTER_POST_ERROR = "REGISTER_POST_ERROR",
  // input
  SAVE_INPUT_USERNAME = "SAVE_INPUT_USERNAME",
  SAVE_INPUT_PASSWORD = "SAVE_INPUT_PASSWORD",
  SAVE_INPUT_EMAIL = "SAVE_INPUT_EMAIL",
  SAVE_INPUT_GENDER = "SAVE_INPUT_GENDER",
  SAVE_INPUT_FILE = "SAVE_INPUT_FILE",
  INCREMENT_PAGE_NUMBER = "INCREMENT_PAGE_NUMBER",
  CHAT_INVITER = "CHAT_INVITER",
  REMEMBER_ME_CHECK = "REMEMBER_ME_CHECK",
  QUICK_LOGIN = "QUICK_LOGIN",
  SIGN_IN = "SIGN_IN",
  SIGN_OUT = "SIGN_OUT",
  TOGGLE_CREATE_GROUP = "TOGGLE_CREATE_GROUP",
  SET_RECIEVER = "SET_RECIEVER",
  SET_FRIEND_REQUEST = "SET_FRIEND_REQUEST",
  SET_USER_SETTINGS = "SET_USER_SETTINGS",
  SET_MOBILE_NAV = "SET_MOBILE_NAV",
  SET_IS_MATCH = "SET_IS_MATCH",
  SET_CHAT_SETTINGS = "SET_CHAT_SETTINGS",
  SET_MODAL_INVITE = "SET_MODAL_INVITE",
  NOTIFICATION_NUMBER = "NOTIFICATION_NUMBER",
  SET_IS_LOGEDIN = "SET_IS_LOGEDIN",
  MESSAGES = "MESSAGES",
  CLEAR = "CLEAR",
  DELETE_MESSAGE = "DELETE_MESSAGE",
  RESET_MESSAGES = "RESET_MESSAGES",
  PAGGINATION_MESSAGES = "PAGGINATION_MESSAGES",
  SHOW_SETTINGS = "SHOW_SETTINGS",
  SET_IS_LOADING = "SET_IS_LOADING",
  SET_WS_CONNECTED = "SET_WS_CONNECTED",
}